requirements.txt
defusedxml==0.7.1
ezdxf==1.1.0
fonttools==4.54.1
fpdf2==2.8.1
pillow==10.4.0
